LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select-in-page/select-in-page-typeahead.tentative-expected.txt



FAIL <select>: Typeahead should focus options without checking them. assert_equals: appearance: base-select must be supported to run this test. expected "base-select" but got "auto"
FAIL <select multiple>: Typeahead should focus options without checking them. assert_equals: appearance: base-select must be supported to run this test. expected "base-select" but got "auto"
